+++ apps/web/src/components/Profile/NftGallery/ReviewSelection.tsx
@@ -1,4 +1,4 @@
-import SingleNft from '@components/NFT/SingleNfts';
+import SingleNft from '@components/NFT/SingleNft';
 import { EmptyState } from '@components/UI/EmptyState';
 import { CollectionIcon, XIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
